@@ -20,56 +20,29 @@
     "keywords": [],
     "license": "MIT",
     "dependencies": {
-        "@madaket/api-token-generator": "^1.1.0",
-        "@vkontakte/icons": "2.7.0",
-        "@vkontakte/vk-bridge": "latest",
-        "@vkontakte/vk-miniapps-deploy": "^0.0.25",
+        "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
+        "@vkontakte/icons": "^2.95.0",
+        "@vkontakte/vk-bridge": "^2.13.0",
+        "@vkontakte/vk-bridge-react": "^1.0.1",
+        "@vkontakte/vk-miniapps-deploy": "^0.1.6",
         "@vkontakte/vk-tunnel": "^0.1.4",
-        "@vkontakte/vkui": "5.1.0",
-        "@vkontakte/vkui-connect": "^1.5.4",
-        "assert": "^2.1.0",
-        "aws-sdk": "^2.1540.0",
-        "axios": "^1.6.1",
+        "@vkontakte/vkui": "5.10.0",
         "babel-eslint": "^10.1.0",
-        "browserify-zlib": "^0.2.0",
         "buffer": "^6.0.3",
-        "child_process": "^1.0.2",
-        "constants-browserify": "^1.0.0",
-        "crypto": "^1.0.1",
         "crypto-browserify": "^3.12.0",
-        "eruda": "^2.4.1",
+        "eruda": "^3.0.1",
         "eruda-code": "^2.0.0",
         "eruda-dom": "^2.0.0",
-        "fs": "^0.0.1-security",
-        "google-protobuf": "^3.21.2",
-        "grpc": "^1.24.11",
-        "html-loader": "^5.0.0",
-        "https-browserify": "^1.0.0",
-        "jsonwebtoken": "^9.0.2",
-        "mock-aws-s3": "^4.0.2",
-        "net": "^1.0.2",
-        "nock": "^13.5.0",
-        "os": "^0.1.2",
-        "os-browserify": "^0.3.0",
-        "path": "^0.12.7",
-        "path-browserify": "^1.0.1",
+        "events": "^3.3.0",
         "prop-types": "^15.7.2",
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2",
-        "redis": "^4.6.10",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
         "stream": "^0.0.2",
-        "stream-browserify": "^3.0.0",
-        "stream-http": "^3.2.0",
-        "timers-browserify": "^2.0.12",
-        "tls": "^0.0.1",
-        "url": "^0.11.3",
-        "util": "^0.12.5"
+        "stream-browserify": "^3.0.0"
     },
     "devDependencies": {
         "cross-env": "^7.0.3",
-        "react-hot-loader": "^4.13.0",
-        "react-scripts": "^5.0.1",
-        "@babel/plugin-proposal-private-property-in-object": "7.21.0-placeholder-for-preset-env.2"
+        "react-scripts": "^5.0.1"
     },
     "resolutions": {
         "react-scripts/webpack-dev-server/yargs/yargs-parser": ">=18.1.2"
