@@ -6,9 +6,12 @@ import config from "../other/config.json";
 
 export const Information = () => {
 	return (
-		<Group separator='hide' mode='plain'>
+		<Group  separator="hide" mode='card' style={{
+			margin: '0 var(--vkui--size_base_padding_horizontal--regular)',
+			textAlign: "justify",
+			padding: 'var(--vkui--size_base_padding_vertical--regular) var(--vkui--size_base_padding_horizontal--regular) '
+		}}>
 			<div style={{
-				margin: '0 var(--vkui--size_base_padding_horizontal--regular)',
 				borderRadius: 'var(--vkui--size_border_radius_paper--regular)',
 				backgroundColor: 'var(--vkui--color_background_content)'
 			}}>
@@ -30,26 +33,24 @@ export const Information = () => {
 					</Paragraph>
 				</div>
 				<div style={{marginBottom: '10px', textAlign: "justify"}}>
-					<Paragraph>
+					<div>
 						<Link href={config.group.href} target="_blank">
-							Разработчики: {config.group.name}
-							<Icon24ExternalLinkOutline width={16} height={16}/>
-						</Link> стремятся обеспечить максимальную точность
-						информации, но в случае обнаружения ошибок или неточностей, они готовы к оперативной их коррекции.
+							Разработчики: {config.group.name} <Icon24ExternalLinkOutline width={16} height={16}/>
+						</Link> стремятся обеспечить максимальную точность информации,
+						но в случае обнаружения ошибок или неточностей, они готовы к оперативной их коррекции.
 						Мы ценим взаимодействие с нашими пользователями и приглашаем вас сообщать о любых замеченных
-						ошибках, предлагать улучшения и делиться вашим опытом. С вашей помощью мы сможем сделать наше приложение
-						ещё лучше!
-					</Paragraph>
+						ошибках, предлагать улучшения и делиться вашим опытом. С вашей помощью мы сможем сделать наше
+						приложение ещё лучше!
+					</div>
 				</div>
 				<div style={{textAlign: "justify"}}>
-					<Paragraph>
+					<div>
 						<Link href='https://hmtpk.ru/' target="_blank">
-							Официальный сайт Ханты-Мансийского технолого-педагогического колледжа
-							<Icon24ExternalLinkOutline width={16} height={16}/>
+							Официальный сайт Ханты-Мансийского технолого-педагогического колледжа <Icon24ExternalLinkOutline width={16} height={16}/>
 						</Link> является официальным источником информации о расписании,
 						новостях и других официальных данных. Рекомендуется обращаться к этому
 						официальному ресурсу для получения достоверной и актуальной информации.
-					</Paragraph>
+					</div>
 				</div>
 			</div>
 		</Group>
