@@ -1,21 +1,36 @@
-import {CardGrid, CustomSelect, Epic, FormItem, FormStatus, Group, Snackbar} from "@vkontakte/vkui";
+import {
+    Button,
+    CardGrid,
+    CustomSelect,
+    Epic,
+    FormItem,
+    FormStatus,
+    Group, Link, Paragraph,
+    Snackbar, Tooltip, TooltipContainer
+} from "@vkontakte/vkui";
 import React, {useEffect} from "react";
 import bridge from "@vkontakte/vk-bridge";
 import ReactDOM from "react-dom";
-import {fetchGroupOrTeacher} from "../other/other";
-import {Icon28CheckCircleOutline, Icon28ErrorCircleOutline} from "@vkontakte/icons";
+import {fetchGroupOrTeacher, updateTooltips} from "../other/other";
+import {Icon24ExternalLinkOutline, Icon28ErrorCircleOutline} from "@vkontakte/icons";
 import config from "../other/config.json";
 
-export const Settings = () => {
-    const [snackbar, setSnackbar] = React.useState(null);
-    const openSuccess = () => {
-        if (snackbar) return;
-        setSnackbar(<Snackbar onClose={() => setSnackbar(null)}
-                              before={<Icon28CheckCircleOutline fill="var(--vkui--color_icon_positive)" />}
-            >Настройки изменены</Snackbar>
-        );
+export const Settings = ({setDisabledExitButton}) => {
+    useEffect(() => {
+        window['groupOrTeacherTemp'] = window['groupOrTeacher']
+    }, [])
+
+    const [groupOrTeacherTemp, setGroupOrTeacherTemp] = React.useState(() => window['groupOrTeacher'])
+    const [tooltip1, setTooltip1] = React.useState( () => window["tooltips"][1]);
+    const [tooltip2, setTooltip2] = React.useState(() => window["tooltips"][2]);
+    const [tooltip3, setTooltip3] = React.useState(() => window["tooltips"][3]);
+    const update = (tooltip, setTooltip) => {
+        window["tooltips"][tooltip] = false
+        setTooltip(false)
+        updateTooltips().then().catch(err => {console.error(err)})
     }
 
+    const [snackbar, setSnackbar] = React.useState(<div></div>);
     const openError = () => {
         if (snackbar) return;
         setSnackbar(<Snackbar onClose={() => setSnackbar(null)}
@@ -24,37 +39,15 @@ export const Settings = () => {
         );
     };
 
-    const storageSet = () => {
-        if (window['userID'] !== 0) {
-            bridge.send(
-                "VKWebAppCallAPIMethod",
-                {
-                    "method": "storage.set",
-                    "params": {
-                        "v": "5.154",
-                        "access_token": config.token,
-                        "key": "schedule",
-                        "user_id": window['userID'],
-                        "value": window['groupOrTeacher']
-                    }
-                }
-            ).then(null).catch((error) => {
-                console.log(error)
-                openError()
-            });
-        } else {
-            ReactDOM.render(<FormStatus mode='error' header='Произошла ошибка' style={{
-                margin: '2px 4px', padding: '5px', justifyContent: 'center', alignItems: 'center'
-            }}>Произошла ошибка при получении Вашего идентификатора
-            </FormStatus>, document.getElementById('settings-menu'))
-        }
-    }
+    const updateGroupOrTeacherTemp = (e) => {
+        window['groupOrTeacherTemp'] = e
+        setGroupOrTeacherTemp(e)
+    };
 
     let [group, setGroup] = React.useState("")
     const onGroupChange = (e) => {
-        window['groupOrTeacher'] = {"group": e.target.value, "teacher": ""}
-        storageSet()
-        openSuccess()
+        updateGroupOrTeacherTemp({"group": e.target.value, "teacher": ""})
+        setDisabledExitButton(false)
         setGroup(e.target.value);
         setTeacher(undefined);
     };
@@ -79,9 +72,8 @@ export const Settings = () => {
 
     let [teacher, setTeacher] = React.useState("");
     const onTeacherChange = (e) => {
-        window['groupOrTeacher'] = {"group": "", "teacher": e.target.value}
-        storageSet()
-        openSuccess()
+        updateGroupOrTeacherTemp({"group": "", "teacher": e.target.value})
+        setDisabledExitButton(false)
         setTeacher(e.target.value);
         setGroup(undefined);
     };
@@ -104,18 +96,20 @@ export const Settings = () => {
         }
     };
 
-    let [menu, setMenu] = React.useState("")
+    let [menu, setMenu] = React.useState("none")
     const onMenuChange = (e) => {setMenu(e.target.value)}
 
     useEffect(() => {
         fetchGroupOrTeacher().then(_ => {
-            if (window['groupOrTeacher'] !== null) {
-                if (window['groupOrTeacher']['group'] !== "") {
-                    setGroup(window['groupOrTeacher']['group'])
+            window['groupOrTeacherTemp'] = window['groupOrTeacher']
+            setGroupOrTeacherTemp(window['groupOrTeacherTemp'])
+            if (window['groupOrTeacherTemp'] !== null) {
+                if (window['groupOrTeacherTemp']['group'] !== "") {
+                    setGroup(window['groupOrTeacherTemp']['group'])
                     setMenu(groupName)
                     fetchGroups()
-                } else if (window['groupOrTeacher']['teacher'] !== "") {
-                    setTeacher(window['groupOrTeacher']['teacher'])
+                } else if (window['groupOrTeacherTemp']['teacher'] !== "") {
+                    setTeacher(window['groupOrTeacherTemp']['teacher'])
                     setMenu(teacherName)
                     fetchTeachers()
                 }
@@ -127,41 +121,112 @@ export const Settings = () => {
             }}>Произошла ошибка при получении Вашей группы
             </FormStatus>, document.getElementById('settings-menu'))
         })
-    }, [null]);
+    }, []);
 
     const teacherName = "Преподаватель"
-    const groupName = "Группа"
+    const groupName = "Студент"
 
-    return <Group id='settings-menu' separator='hide' mode='plain'>
-        <CardGrid size="l">
-            <FormItem
+    return <Group id='settings-menu' separator='hide' mode='plain' style={{height: 'calc(100vh - var(--vkui--size_panel_header_height--regular)*4)'}}>
+        <CardGrid size="l" style={{height: '100%', display: 'flex', margin: '0', padding: '0'}}>
+            <TooltipContainer
+                fixed
                 style={{
-                    flex: '1', padding: '0 0 var(--vkui--size_base_padding_vertical--regular) 0',
+                    flex: '1',
+                    padding: '0 var(--vkui--size_base_padding_horizontal--regular)'
                 }}>
-                <CustomSelect
-                    placeholder="Выберите тип расписания" value={menu} onChange={onMenuChange}
-                    options={[{label: groupName, value: groupName}, {label: teacherName, value: teacherName}]}
-                />
-            </FormItem>
-            <Epic activeStory={menu} style={{padding: 'var(--vkui--size_base_padding_vertical--regular) 0 0'}}>
-                <Group id={groupName} separator="hide" mode='plain'>
-                    <FormItem
-                        style={{flex: '1', padding: '0'}}>
+                <Tooltip
+                    style={{textAlign: 'center'}}
+                    text="Добро пожаловать в раздел настройки расписания! Здесь вы можете настроить ваш опыт использования, выбрав тип пользователя: 'Студент' или 'Преподаватель'."
+                    isShown={tooltip1} onClose={() => update(1, setTooltip1)}
+                >
+                    <FormItem style={{padding: '0'}}>
                         <CustomSelect
-                            placeholder="Выберите группу" searchable options={groupOptions} onChange={onGroupChange}
-                            value={group} onOpen={groupOptions.length === 0 && fetchGroups} fetching={groupFetching}
+                            placeholder="Выберите тип пользователя" value={menu} onChange={onMenuChange}
+                            options={[{label: groupName, value: groupName}, {label: teacherName, value: teacherName}]}
                         />
                     </FormItem>
+                </Tooltip>
+                <Epic activeStory={menu} style={{padding: 'var(--vkui--size_base_padding_vertical--regular) 0 0'}}>
+                    <Group id='none' separator="hide" mode='plain'>
+                        <FormItem style={{flex: '1', padding: '0'}}>
+                            <CustomSelect disabled
+                                          placeholder="" searchable options={groupOptions} onChange={onGroupChange}
+                                          value={group} onOpen={groupOptions.length === 0 && fetchGroups}
+                                          fetching={groupFetching}
+                            />
+                        </FormItem>
+                    </Group>
+                    <Group id={groupName} separator="hide" mode='plain'>
+                        <Tooltip
+                            style={{textAlign: 'center'}}
+                            text="Укажите свою группу из предложенного списка. Это позволит вам видеть расписание, соответствующее вашей программе обучения."
+                            isShown={tooltip2 && !tooltip1} onClose={() => {
+                            update(2, setTooltip2)
+                            if (groupOrTeacherTemp['group'] !== "" || groupOrTeacherTemp['teacher'] !== "") {
+                                setDisabledExitButton(false)
+                            }
+                        }}
+                        >
+                            <FormItem style={{padding: '0'}}>
+                                <CustomSelect
+                                    placeholder="Выберите группу" searchable options={groupOptions}
+                                    onChange={onGroupChange}
+                                    value={group} onOpen={groupOptions.length === 0 && fetchGroups}
+                                    fetching={groupFetching}
+                                />
+                            </FormItem>
+                        </Tooltip>
+                    </Group>
+                    <Group id={teacherName} separator="hide" mode='plain'>
+                        <Tooltip
+                            style={{textAlign: 'center'}}
+                            text="Укажите своё ФИО. Таким образом, вы получите расписание, соответствующее вашим учебным занятиям."
+                            isShown={tooltip3 && !tooltip1} onClose={() => {
+                            update(3, setTooltip3)
+                            if (groupOrTeacherTemp['group'] !== "" || groupOrTeacherTemp['teacher'] !== "") {
+                                setDisabledExitButton(false)
+                            }
+                        }}
+                        >
+                            <FormItem style={{padding: '0'}}>
+                                <CustomSelect
+                                    placeholder="Выберите преподавателя" searchable options={teacherOptions}
+                                    onChange={onTeacherChange}
+                                    value={teacher} onOpen={teacherOptions.length === 0 && fetchTeachers}
+                                    fetching={teacherFetching}
+                                />
+                            </FormItem>
+                        </Tooltip>
+                    </Group>
+                </Epic>
+                <Group separator="hide" mode='card' style={{
+                    margin: '10px 0', textAlign: "justify",
+                    padding: 'var(--vkui--size_base_padding_vertical--regular) var(--vkui--size_base_padding_horizontal--regular) '
+                }}>
+                    <div style={{marginBottom: '10px'}}>
+                        <Paragraph>
+                            Если вы не нашли нужную группу или преподавателя в списке, пожалуйста,
+                            сообщите об этом разработчикам. Для отправки обратной связи и уведомления об
+                            отсутствующих данных, перейдите по следующей ссылке:
+                            <Link href={config.group.href} target="_blank"> Разработчики: {config.group.name}
+                                <Icon24ExternalLinkOutline width={16} height={16}/></Link>.
+                        </Paragraph>
+                    </div>
+                    <div>
+                        <Paragraph>
+                            Благодарим за ваше внимание и помощь в совершенствовании нашего сервиса!
+                        </Paragraph>
+                    </div>
                 </Group>
-                <Group id={teacherName} separator="hide" mode='plain'>
-                    <FormItem style={{flex: '1', padding: '0'}}>
-                        <CustomSelect
-                            placeholder="Выберите преподавателя" searchable options={teacherOptions} onChange={onTeacherChange}
-                            value={teacher} onOpen={teacherOptions.length === 0 && fetchTeachers} fetching={teacherFetching}
-                        />
-                    </FormItem>
-                </Group>
-            </Epic>
+                <div hidden={userID !== 390295814}>
+                    <Button appearance='negative' align="center" mode="outline" stretched={true} onClick={() => {
+                        updateGroupOrTeacherTemp({"group": "", "teacher": ""})
+                    }}
+                            style={{margin: 'calc(var(--vkui--size_base_padding_vertical--regular)*1) 0 0'}}
+                            disabled={groupOrTeacherTemp['group'] === "" && groupOrTeacherTemp['teacher'] === ""}
+                    >Очистить</Button>
+                </div>
+            </TooltipContainer>
             {snackbar}
         </CardGrid>
     </Group>
